app/requirements.txt
Flask==2.0.1
Flask-WTF==1.2.2
Werkzeug==2.0.3
scikit-learn==1.2.2
segment-analytics-python==2.3.3
numpy==1.24.4
sentence-transformers==2.3.1
torch==2.0.1
huggingface_hub==0.15.1
openai==0.28.0
PyYAML==6.0.2
GitPython==3.1.44
redis==6.2.0
fakeredis==2.30.1
